requirements.txt
numpy
pandas
matplotlib
opencv-python
albumentations
ipykernel
tqdm
#pycocotools   #Comment when creating Docker Image
cython
Pillow
torch-summary
fastapi
uvicorn
Jinja2
python-multipart
PyYAML
dill==0.3.5.1
utils
-e .